Provide `TcpListener` to `run`
Binding to the address manually can fail during tests if the application
is already running and has taken port 8000. When using port 0 to let the
OS choose a random free port, the tests no longer know which port to
send requests to. By constructing a `TcpListener` outside the
`spawn_app` function, we can grab the full address, and pass the
`TcpListener` to the `spawn_app` function afterwards.